remove comments, update VS version
