+++ LanMap Filter/LanMap Filter.vcxproj
@@ -94,7 +94,9 @@
     <ClInclude Include="tlvlldp.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="cdp.cpp" />
     <ClCompile Include="LanMap Filter.cpp" />
+    <ClCompile Include="lldp.cpp" />
     <ClCompile Include="stdafx.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
